Move verify command to its own file
